Add toLowerCase func to alter playerChoice string
Before, the function would alter the playChoice string so that
the first character would be uppercase, and the remaining characters
would be unchanged. Now, the playerChoice string will be changed so that
ONLY the first characters are uppercase, and the remaining characters
are all lower case. This will help letter on when comparing the string
in conditionals to check whether the input is valid